# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "This is the API documentation for the Social API project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/followers": {
            "post": {
                "tags": [
                    "Follower"
                ],
                "summary": "Follow a user",
                "description": "Allows a user to follow another user.",
                "operationId": "followUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "follower_id",
                                    "followed_id"
                                ],
                                "properties": {
                                    "follower_id": {
                                        "description": "ID of the profile following the other user",
                                        "type": "integer"
                                    },
                                    "followed_id": {
                                        "description": "ID of the profile being followed",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Follow created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Follow relation ID",
                                            "type": "integer"
                                        },
                                        "follower_id": {
                                            "description": "Follower profile ID",
                                            "type": "integer"
                                        },
                                        "followed_id": {
                                            "description": "Followed profile ID",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Follower"
                ],
                "summary": "Unfollow a user",
                "description": "Allows a user to unfollow another user.",
                "operationId": "unfollowUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "follower_id",
                                    "followed_id"
                                ],
                                "properties": {
                                    "follower_id": {
                                        "description": "ID of the profile unfollowing the other user",
                                        "type": "integer"
                                    },
                                    "followed_id": {
                                        "description": "ID of the profile being unfollowed",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Unfollowed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Success message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Follow relation not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/likes": {
            "post": {
                "tags": [
                    "Like"
                ],
                "summary": "Create a like for a post",
                "description": "Allows a user to like a specific post.",
                "operationId": "likePost",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "post_id",
                                    "profile_id"
                                ],
                                "properties": {
                                    "post_id": {
                                        "description": "ID of the post to like",
                                        "type": "integer"
                                    },
                                    "profile_id": {
                                        "description": "ID of the profile liking the post",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Like created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Like ID",
                                            "type": "integer"
                                        },
                                        "profile_id": {
                                            "description": "Profile ID that liked the post",
                                            "type": "integer"
                                        },
                                        "post_id": {
                                            "description": "Post ID that is liked",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Like"
                ],
                "summary": "Remove a like from a post",
                "description": "Allows a user to remove their like from a specific post.",
                "operationId": "removeLike",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "post_id",
                                    "profile_id"
                                ],
                                "properties": {
                                    "post_id": {
                                        "description": "ID of the post to unlike",
                                        "type": "integer"
                                    },
                                    "profile_id": {
                                        "description": "ID of the profile removing the like",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Like removed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Success message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Like not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your token here",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}